+++ src/main/pageTweaks/index.js
@@ -12,12 +12,12 @@ import performObservation from 'operations/performObservation';
 import { Sentry } from 'plugins/sentry';
 import addOns from 'addons/index';
 import driveManager from 'plugins/driveManager';
-import submenu from 'main/submenu';
 
 import newTrade from 'main/pageTweaks/newTrade';
 import uebersicht from 'main/pageTweaks/uebersicht';
 import produktion from 'main/pageTweaks/produktion';
 import buildingTree from 'main/pageTweaks/buildingTree';
+import inbox from 'main/pageTweaks/inbox';
 
 import moment from 'moment';
 import momentDurationFormatSetup from 'moment-duration-format';
@@ -275,62 +275,7 @@ const pageTweaks = {
       config.loadStates.content = false;
     });
   },
-  inbox: () => {
-    // clear content so loadStates doesn't fire too early
-    // lwmJQ('#inboxContent').html('');
-    config.promises.content = getPromise('.inboxDeleteMessageButtons,#messagesListTableInbox');
-    config.promises.content.then(() => {
-      config.loadStates.content = false;
-
-      // workaround to bring the submenu in if you come to message from anywhere else than the message menu button
-      if (lwmJQ('#veticalLink a.navButton').length !== 0 && lwmJQ('.secound_line a.navButton').length === 0) submenu.move();
-
-      // go through messages and add direct link to fight and spy reports
-      // we do this after updating loadstate to not slow down page load
-      const addReportLink = (message) => {
-        const type = message.subject.search(/Kampfbericht/) !== -1 ? 'view_report_attack' : 'planetenscanner_view';
-        const $link = lwmJQ(`<a target='_blank' href='https://last-war.de/${type}.php?id=${message.reportID}&user=${config.gameData.playerID}'><i style='margin-left: 5px;' class='fas fa-external-link-alt'></i></a>`);
-        lwmJQ(`[onclick*='${message.id}']`).after($link);
-      };
-
-      // install handler to attach report links on browsing message pages
-      if (!config.pages.inbox.reportHandler) {
-        lwmJQ(document).on('click', (e) => {
-          const check = lwmJQ(e.target).is('.formButton[onclick*=\'nextPage\']') || lwmJQ(e.target).is('.formButton[onclick*=\'previousPage\']');
-          if (!check) return;
-          if (![2, 4].includes(siteWindow.window.current_view_type)) return;
-          lwmJQ.each(config.gameData.messageData[1], (i, m) => {
-            if (m.subject.search(/Kampfbericht|Spionagebericht/) !== -1 && m.user_nickname === 'Systemnachricht') addReportLink(m);
-          });
-        });
-        config.pages.inbox.reportHandler = true;
-      }
-
-      if ([2, 4].includes(siteWindow.window.current_view_type) && gmConfig.get('message_spylinks')) {
-        lwmJQ.each(config.gameData.messageData[1], (i, m) => {
-          if (m.subject.search(/Kampfbericht|Spionagebericht/) !== -1 && m.user_nickname === 'Systemnachricht') {
-            siteWindow.jQuery.ajax({
-              url: `/ajax_request/get_message_info.php?id_conversation=${m.id}&current_view_type=${siteWindow.window.current_view_type}`,
-              dataType: 'json',
-              data: { lwm_ignoreProcess: 1 },
-              success(data) {
-                // add reportID to data for future use
-                [, config.gameData.messageData[1][i].reportID] = data[0][0].text.match(/id=(\d*)/);
-
-                addReportLink(config.gameData.messageData[1][i]);
-              },
-              timeout: config.promises.interval.ajaxTimeout,
-            });
-          }
-        });
-      }
-    }).catch((e) => {
-      Sentry.captureException(e);
-      // console.log(e);
-      throwError();
-      config.loadStates.content = false;
-    });
-  },
+  inbox,
   trades: () => {
     // we have to chain promises here to work around an issue
     // we resolve the page on #link since #tradeOfferDiv is not present without any active trades